Rename blank to sparse.
