+++ src/selection/enter.js
@@ -1,5 +1,5 @@
-import blank from "./blank";
+import sparse from "./sparse";
 
 export default function() {
-  return this._enter || ((this._enter = blank(this))._update = this, this._enter);
+  return this._enter || ((this._enter = sparse(this))._update = this, this._enter);
 };
